keep lib_wasm swc agnostic (#105)
* Keep lib_wasm swc agnostic

* Use swc Stmt::is_use_strict
